Revert 'fix(apple): Make Favorites load/save async' (#8068)
Turns out I was wrong about UserDefaults - they're persisted, but backed
by an in-memory cache. The persisting is handled asynchronously, so this
is unlikely to be the culprit of
https://firezone-inc.sentry.io/issues/6225229650/?project=4508175177023488&query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D&referrer=issue-stream&statsPeriod=30d&stream_index=1

More likely it has to do with some other synchronous I/O, so the search
continues. Unfortunately we don't have great backtraces for this issue
because it looks like the hang is happening outside the scope of our
code, and seems to be macOS-only in low memory conditions.

Reverts firezone/firezone#8060