refactor(apple): Collapse ViewModels to app-wide Store (#8218)
Our application state is incredibly simple, only consisting of a handful
of properties.

Throughout our codebase, we use a singular global state store called
`Store`. We then inject this as the singular piece of state into each
view's model.

This is unnecessary boilerplate and leads to lots of duplicated logic.
Instead we refactor away (nearly) all of the application's view models
and instead use an `@EnvironmentObject` to inject the store into each
view.

This convention drastically simplifies state tracking logic and
boilerplate in the views.